Choice of theme complete
For the first stage, I decided to implement the theme from get go because I liked it. But I don't know maybe I'll change it and build my own layout. For now though I have an idea of maybe starting with the dark theme and on toggling the lights to come on the theme changes.