First phase of switching lights on complete
Utilizing some of the fancy nbtoggle features of the theme to help change the theme to show that lights are on moving from dark to default. Next step is to get images and some icons and the animation to fade out the toggle so she doesn't click it to switch off the lights.

Note to self -- At the end of it all just set a "Do It All Over" button that refreshes the page and takes her back to the dark screen mode.