Part 1 of phase 2 complete
Added an image of lights and it's at the top of the column and centered. So need to add animations to load them in. And then test with the next button that will be used to bring in the balloons.

Note to Self: -- Ask someone how this looks. After asking changed image used.