Restructured lights component and Balloons added
There was a restructure for the lights component and change in plans removed use of the picture since moving from dark mode to light mode is already turning lights on.

Balloons have been added animation still looks a bit shaky might leave fine tuning it for later but I like current look of it. Recording was made with OBS to demo progress up until that point.