Added cake and new button and new animation.
Cake component added that brings in nice cake and candle code we found online. Switched to buttons instead of toggles so there was a change in that aspect and then trying to now switch to fade in and out from Cake Button onwards but fade in for cake button being a bit fidgety. Fade out works just fine though. Next up are audio and then a scrolling message.