Added confetti and made some further changes to app
Confetti has been added as a separate component that appears only after cake is in place. In regards to further changes made, I modified the colors being used in the balloons. My string disappeared but for some reason I like how it turned out. Buttons starting from Place Cake need to be fixed since the canvas created to help bring the confetti to life forced the button for placing cake out of position and so it looks very odd.